Introduced new way of handling floating point numbers, as long long. The long long represents the floating point number xe4.

+++ inout.h
@@ -2,5 +2,5 @@
 
 void init(std::string);
 int readStation();
-int readPoint(std::string &, double &, double &);
+int readPoint(std::string &, long long &, long long &);
